requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
numpy==1.25.1
pandas==2.0.3
psycopg==3.1.10
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
XlsxWriter==3.1.6